remove conftest
